refactor: controller, domain, dto 패키지 분리
