+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,8 +1,10 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationResponse;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
